moves modal portal to ApplicationRoot
positions the modal container properly
