+++ client/components/ApplicationRoot.js
@@ -1,5 +1,6 @@
 //@flow
 import * as React from 'react';
+import ReactDOM from 'react-dom';
 import { ApolloProvider as ApolloHooksProvider } from 'react-apollo-hooks';
 import urljoin from 'url-join';
 import ApolloClient from 'apollo-client';
@@ -15,6 +16,57 @@ import { useToastState, ToasterContext } from 'global/hooks/toaster';
 import { css } from 'uikit';
 import ToastStack from 'uikit/notifications/ToastStack';
 import { AuthProvider } from 'global/hooks/useAuthContext';
+import Modal from 'uikit/Modal';
+
+const modalPortalRef = React.createRef();
+const useMounted = () => {
+  const [mounted, setMounted] = React.useState(false);
+  React.useEffect(() => {
+    setMounted(true);
+  }, []);
+  return mounted;
+};
+export const ModalPortal = ({ children }: { children: React.Node }) => {
+  const ref = modalPortalRef.current;
+  const mounted = useMounted();
+  return ref
+    ? ReactDOM.createPortal(
+        <div
+          id="modalContainer"
+          css={css`
+            transition: all 0.2s;
+            height: 100vh;
+            width: 100vw;
+            opacity: ${mounted ? 1 : 0};
+          `}
+        >
+          <Modal.Overlay>{children}</Modal.Overlay>
+        </div>,
+        ref,
+      )
+    : null;
+};
+
+const ToastDisplayArea = ({ toaster }: { toaster: $Call<typeof useToastState, void> }) => (
+  <div
+    className="toastStackContainer"
+    css={css`
+      position: fixed;
+      z-index: 9999;
+      right: 0px;
+      top: 80px;
+    `}
+  >
+    <div
+      css={css`
+        margin-right: 20px;
+        margin-left: 20px;
+      `}
+    >
+      <ToastStack toastConfigs={toaster.toastStack} onInteraction={toaster.onInteraction} />
+    </div>
+  </div>
+);
 
 export default function ApplicationRoot({
   egoJwt,
@@ -71,27 +123,16 @@ export default function ApplicationRoot({
                 <ThemeProvider>
                   <>
                     {children}
+                    <ToastDisplayArea toaster={toaster} />
                     <div
-                      className="toastStackContainer"
                       css={css`
                         position: fixed;
+                        left: 0px;
+                        top: 0px;
                         z-index: 9999;
-                        right: 0px;
-                        top: 80px;
                       `}
-                    >
-                      <div
-                        css={css`
-                          margin-right: 20px;
-                          margin-left: 20px;
-                        `}
-                      >
-                        <ToastStack
-                          toastConfigs={toaster.toastStack}
-                          onInteraction={toaster.onInteraction}
-                        />
-                      </div>
-                    </div>
+                      ref={modalPortalRef}
+                    />
                   </>
                 </ThemeProvider>
               </ApolloHooksProvider>
